files: split Style.go into StyleIDs.go
it was pretty hard to find appropiate styleID in that large style.go file - i moved ID stuff into StyleIDs.go
